third_party/blink/web_tests/css-parser/deprecated-css-appearance-values-warning/appearance-with-nonstandard-keywords-expected.txt

CONSOLE WARNING: The keyword 'push-button' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'square-button' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'inner-spin-button' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-slider' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-sliderthumb' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-volume-slider' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'media-volume-sliderthumb' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'slider-horizontal' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'slider-vertical' specified to an 'appearance' property is not standardized. It will be removed in the future and replaced by vertical writing-mode (currently experimental).
CONSOLE WARNING: The keyword 'sliderthumb-horizontal' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'sliderthumb-vertical' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
CONSOLE WARNING: The keyword 'searchfield-cancel-button' used on the 'appearance' property was deprecated and has now been removed. It will no longer have any effect.
This is a testharness.js-based test.
[FAIL] There should be console warnings about non-standard appearance keywords
  assert_true: This should FAIL so we generate a -expected.txt expected true got false
Harness: the test ran to completion.
